requirements.txt
Django==3.2.25
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.7
python-dotenv==1.0.1
strawberry-graphql==0.229.2
django-cacheops


